+++ src/main/java/calculator/Calculator.java
@@ -1,45 +1,33 @@
-package Calculator;
+package calculator;
 
 import java.util.Scanner;
 
-public class MainTest2 {
+public class Calculator {
 
-    public static void main(String[] args) {
-        System.out.println("문자열을 입력 하세요");
+    static final String STRING_BLANK = " ";
 
-        int result = makeResult(input());
-
-        System.out.println(result);
-
-    }
-
-    // 1. 문자열 입력
     static String input() {
         Scanner sc = new Scanner(System.in);
         return sc.nextLine();
     }
 
-    // 2. 입력값 공백/null 체크
     static boolean isBlank(String input) {
-        if (input.equals(" ") || input == null) {
-            return true;
+        if (input == null ||input.equals(STRING_BLANK) || input.isEmpty()) {
+            throw new IllegalArgumentException("IllegalArgumentException !!!");
         }
         return false;
     }
 
-    // 2-1. 입력값 공백 제거
     static String[] splitBlank(String input) {
-        return input.split(" ");
+        return input.split(STRING_BLANK);
     }
 
-    // 3. 입력 값 실행
     static int makeResult(String input) {
         if (isBlank(input))
-            throw new RuntimeException();
+            throw new IllegalArgumentException("IllegalArgumentException !!!");
         return calculateSplitedString(splitBlank(input));
     }
 
-    // 4. 입력 값 계산 실행
     static int calculateSplitedString(String[] str) {
         int result = toInt(str[0]);
         for (int i = 0; i < str.length - 2; i += 2) {
@@ -57,7 +45,7 @@ public class MainTest2 {
             return multiplication(first, second);
         if (operator == '/')
             return division(first, second);
-        throw new RuntimeException();
+        throw new IllegalArgumentException("IllegalArgumentException !!!");
     }
 
     static int addition(int i, int j) {
